:construction_worker: [ci] temp fix race tests on go 1.14
This seems to be a bug in go's crypto lib, which is used by go-ethereum
https://github.com/ethereum/go-ethereum/issues/20731

From the Go 1.14 release notes:

This release adds -d=checkptr as a compile-time option for adding
instrumentation to check that Go code is following unsafe.Pointer safety
rules dynamically. This option is enabled by default (except on Windows)
with the -race or -msan flags, and can be disabled with
-gcflags=all=-d=checkptr=0. Specifically, -d=checkptr checks the
following:

When converting unsafe.Pointer to *T, the resulting pointer must be
aligned appropriately for T.
If the result of pointer arithmetic points into a Go heap object, one of
the unsafe.Pointer-typed operands must point into the same object.

Using -d=checkptr is not currently recommended on Windows because it
causes false alerts in the standard library.